% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cocoon_open}
\alias{cocoon_open}
\title{Load socials data from local file}
\usage{
cocoon_open()
}
\value{
Socials data frame
}
\description{
Reads your local socials data for perusal or modification.
}
\examples{
\dontshow{if (local_eg()) withAutoprint(\{ # examplesIf}
cocoon_open()
\dontshow{\}) # examplesIf}
}
