% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cocoon_update}
\alias{cocoon_update}
\title{Save/Update socials data to local file}
\usage{
cocoon_update(socials, type = NULL, value = NULL)
}
\arguments{
\item{socials}{Data frame. Data frame of previously fetched/loaded social
contact information. Data frame with three columns `type`, `value` and
`github` to indicate the type and value of the contact information and the
GitHub username identifying the individual.}

\item{type}{Character. Type of `value` (e.g., "github" for github handle)}

\item{value}{Character. Value to search by}
}
\value{
Nothing
}
\description{
Save/Update your local copy of the contact information.
}
\examples{
\dontshow{if (local_eg()) withAutoprint(\{ # examplesIf}
socials_fetch("steffilazerte") |>
  cocoon_update()

# Or directly
cocoon_update("steffilazerte")

# Or add parts
cocoon_update("steffilazerte", type = "name", value = "Stefanie LaZerte")
\dontshow{\}) # examplesIf}
}
