% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch-package.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{socials}{Data frame. Data frame of previously fetched/loaded social
contact information. Data frame with three columns `type`, `value` and
`github` to indicate the type and value of the contact information and the
GitHub username identifying the individual.}

\item{github}{Character. Github username}

\item{names}{Character. Names to fetch by.}

\item{value}{Character. Value to search by}

\item{type}{Character. Type of `value` (e.g., "github" for github handle)}

\item{pkg}{Character. (Optional) Repository name (package name).}

\item{owner}{Character. (Optional) Owner of the repository.}

\item{force_masto}{Logical. Whether to force a re-fetching of Mastodon handles.}

\item{open_browser}{Logical. Whether to open the profile page in a browser
for confirmation.}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use `@inheritParams common_docs` to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
