% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{email_from_universe}
\alias{email_from_universe}
\title{Get email from R-universe}
\usage{
email_from_universe(name, package = NULL, universe = "ropensci-staging")
}
\arguments{
\item{name}{Name of the package contributor}

\item{package}{Name of the package. If not present, we'll look for any person
with that name in the R-universe.}

\item{universe}{Name of the R-universe.}
}
\value{
An email
}
\description{
Get email from R-universe
}
\examples{
email_from_universe("Maëlle Salmon", universe = "ropensci")
email_from_universe("Maelle S Salmon", universe = "ropensci")
}
