% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch a value}
\usage{
fetch(values, type = "mastodon", n = 1)
}
\arguments{
\item{values}{Character. Values of social data to match by.}

\item{type}{Character. Type of social data to fetch.}

\item{n}{Numeric. Maximum number of matches per value to return.}
}
\value{
Character string of the value of the type of data requested
}
\description{
Fetches a single value from your local socials data frame.
}
\examples{
\dontshow{if (local_eg()) withAutoprint(\{ # examplesIf}
fetch("Steffi LaZerte", type = "linkedin")
fetch(c("Steffi LaZerte", NA, "Yanina Bellini Saibene"))
fetch("steffilazerte")
fetch("Maëlle Salmon", type = "github")
\dontshow{\}) # examplesIf}
}
