% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{fmt_masto}
\alias{fmt_masto}
\title{Convert a mastodon user link to handle}
\usage{
fmt_masto(x)
}
\arguments{
\item{x}{Character. Link to user's profile}
}
\value{
Character user handle @user@instance
}
\description{
Convert a mastodon user link to handle
}
\examples{
\dontshow{if (local_eg()) withAutoprint(\{ # examplesIf}
fmt_masto("https://fosstodon.org/@steffilazerte")
fmt_masto("steffi lazerte")
fmt_masto("@steffilazerte@fosstodon.org")
fmt_masto(NA)
fmt_masto(c("https://fosstodon.org/@steffilazerte", "https://hackyderm.io/@ropensci"))
fmt_masto("none")
\dontshow{\}) # examplesIf}
}
