% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch-package.R
\docType{package}
\name{monarch-package}
\alias{monarch}
\alias{monarch-package}
\title{monarch: Manage rOpenSci Social Contacts}
\description{
Tools to find and keep track of contact info and social media handles of rOpenSci community members.
}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})

}
\keyword{internal}
