% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{socials_fetch}
\alias{socials_fetch}
\title{Fetch contact information from social media}
\usage{
socials_fetch(
  github = NULL,
  name = NULL,
  pkg = NULL,
  owner = "ropensci",
  which_cols = c("github", "name", "mastodon"),
  force_masto = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{github}{Character. Github username}

\item{name}{Character. Alternatively, a full name (can be slow if `pkg` not
also provided).}

\item{pkg}{Character. Repository name for an rOpenSci package (in the
`ropensci` organization)}

\item{owner}{Character. (Optional) Owner of the repository.}

\item{which_cols}{Character vector. Social handles to include defaults to
"github", "name", "mastodon".}

\item{force_masto}{Logical. Whether to force a re-fetching of Mastodon handles.}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
Data frame of contact information
}
\description{
This is a wrapper around the three `socials_xxx()` functions. `socials_gh()`
starts (ideally) with a GitHub username to fetch contact info from GitHub,
followed by `socials_ro()` which checks rOpenSci author pages and finally, if
Mastodon handles have not been found `socials_masto()`.
}
\details{
This workflow generally assumes that the person has a GitHub username. You
can supply name, which will trigger a search for the GitHub username for you.
This will search all of the rOpenSci organization for users with that name
which can take a while. However, if you supply a name and a `pkg` (repository
in rOpenSci), it can be much faster.
}
\examples{
\dontshow{if (local_eg()) withAutoprint(\{ # examplesIf}
socials_fetch("steffilazerte")
socials_fetch(name = "Steffi LaZerte", pkg = "weathercan")
socials_fetch(name = "Bart Vanhoorne", pkg = "worrms")
socials_fetch(name = "Maelle Salmon", pkg = "babelquarto", owner = "ropensci-review-tools")
\dontshow{\}) # examplesIf}
}
