% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{socials_masto}
\alias{socials_masto}
\title{Fetch Mastodon handle}
\usage{
socials_masto(
  socials = NULL,
  names = NULL,
  github = NULL,
  force = FALSE,
  open_browser = interactive(),
  quiet = FALSE
)
}
\arguments{
\item{socials}{Data frame. Data frame of previously fetched/loaded social
contact information. Data frame with three columns `type`, `value` and
`github` to indicate the type and value of the contact information and the
GitHub username identifying the individual.}

\item{names}{Character. Names to fetch by.}

\item{github}{Character. Github username}

\item{force}{Logical. Whether to force a search even if the handle already
exists.}

\item{open_browser}{Logical. Whether to open the profile page in a browser
for confirmation.}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
If `socials` or `github` are provided, a data frame of social contact
  information. Three columns `type`, `value` and `github` to indicate the
  type and value of the contact information and the GitHub username
  identifying the individual. Otherwise a character string of the mastodon
  handle.
}
\description{
Useful if neither GitHub nor the rOpenSci author pages contain a Mastodon
handle. Because this can take a bit of time, if a mastodon handle is
detected in the `socials` data frame, `socials_masto()` will skip the search,
unless `force = TRUE`. Useful if you think the handle is incorrect.
}
\details{
Unlike most other functions, `socials_masto()` permits searching for a handle
by name without providing a `socials` data frame or a `github` handle. In
this case, the Mastodon handle is returned as a character string and not as
a socials data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
socials_gh("steffilazerte") |>
 socials_masto(force = TRUE)

socials_masto(github = "steffilazerte")
socials_masto(names = "steffi")

socials_masto(names = "steffi", github = "steffilazerte")
\dontshow{\}) # examplesIf}
}
