% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_handles}
\alias{add_handles}
\title{Find and insert missing social media handles}
\usage{
add_handles(df, primary = "github", pkg_col = NULL, prefix = NULL)
}
\arguments{
\item{df}{Data frame with columns appended with `_name`, `_github`, `_mastodon`.}
}
\value{
Data frame with added names and social media handles.
}
\description{
Find and insert missing social media handles
}
\examples{
d <- data.frame(author_name = "Steffi LaZerte")
add_handles(d, primary = "name", prefix = "author_")

d <- data.frame(github = "steffilazerte")
add_handles(d)
}
