% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cache_check}
\alias{cache_check}
\title{Check Cached Dir}
\usage{
cache_check()
}
\value{
Logical. TRUE if it exists, FALSE if not.
}
\description{
Checks for the presence of a cache dir. If interactive and not present,
asks user if it should be created. If non-interactive and not present, it
is created automatically.
}
\examples{
cache_check()
}
