% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cocoon_fetch}
\alias{cocoon_fetch}
\title{Fetch all details on a social contact}
\usage{
cocoon_fetch(value, type = "github")
}
\arguments{
\item{value}{Character. Value to search by}

\item{type}{Character. Type of `value` (e.g., "github" for github handle)}
}
\description{
Fetch all details on a social contact
}
\examples{
cocoon_fetch("steffilazerte")
cocoon_fetch("Steffi LaZerte", type = "name")
}
