% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cocoon_remove}
\alias{cocoon_remove}
\title{Remove Socials by GitHub handle}
\usage{
cocoon_remove(github)
}
\arguments{
\item{github}{Character. Github username}
}
\value{
Socials data frame
}
\description{
Remove Socials by GitHub handle
}
\examples{
cocoon_fetch("steffilazerte")
cocoon_remove("steffilazerte")
cocoon_fetch("steffilazerte")
}
