% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{cocoon_update}
\alias{cocoon_update}
\title{Save/Update socials data to local file}
\usage{
cocoon_update(socials, type = NULL, value = NULL)
}
\value{
Nothing
}
\description{
Save/Update your local copy of the contact information.
}
\examples{
socials_fetch("steffilazerte") |>
  cocoon_update()

# Or directly
cocoon_update("steffilazerte")

# Or add parts
cocoon_update("steffilazerte", type = "name", value = "Stefanie LaZerte")

}
