% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{gh_search}
\alias{gh_search}
\title{Find GH username from repository and full name}
\usage{
gh_search(name, pkg = NULL, owner = "ropensci", open_browser = TRUE)
}
\arguments{
\item{name}{Character. Full or partial name of the person for whom you want
to fetch the GitHub username.}

\item{pkg}{Character. (Optional) Repository name (package name).}

\item{owner}{Character. (Optional) Owner of the repository.}

\item{open_browser}{Logical. Whether to open the profile page in a browser
for confirmation.}
}
\value{
Accepted user name
}
\description{
Look up users of a repository and match to a name. Try with and without
initials.
}
\examples{

gh_user(name = "Steffi E. LaZerte", owner = "ropensci", pkg = "weathercan")
gh_user(name = "Steffi", owner = "ropensci", pkg = "weathercan")
}
