% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch-package.R
\docType{package}
\name{monarch-package}
\alias{monarch}
\alias{monarch-package}
\title{monarch: Manage rOpenSci Social Contacts}
\description{
Tools to find and keep track of contact info and social media handles of rOpenSci community members.
}
\keyword{internal}
