% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{socials_gh}
\alias{socials_gh}
\title{Fetch contact information from GitHub}
\usage{
socials_gh(github, quiet = FALSE)
}
\arguments{
\item{github}{Character. Github username}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
Data frame of social contact information. Three columns `type`,
  `value` and `github` to indicate the type and value of the contact
  information and the GitHub username identifying the individual.
}
\description{
Using a GitHub username fetch the contact information from their profile.
}
\examples{

socials_gh("steffilazerte")
}
